pyproject.toml
[tool.poetry]
name = "bookstore1"
version = "0.1.0"
description = "Bookstore1 API"
authors = ["Nunes"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
